+++ src/index.d.ts
@@ -1,6 +1,21 @@
 /** @format */
 
+/**
+ * Types
+ */
+declare type ipcReplyTypes = "version" | "data" | "update";
+declare type ipcTypes = "main" | "background";
 declare interface ipcReply {
-  type: "version" | "data" | "update";
-  payload: any
+  type: ipcReplyTypes;
+  payload?: any;
+}
+
+declare interface ipc {
+  event: string;
+  channel: ipcTypes;
+}
+
+declare interface ipcPayload {
+  event: string;
+  payload?: any;
 }